requirements.txt
asgiref==3.8.1
Django==5.0.6
django-quill==0.1.0
django-quill-editor==0.1.40
django-rest-framework==0.1.0
djangorestframework==3.15.1
sqlparse==0.5.0
typing_extensions==4.11.0
